==> 6 - TABELAS_DE_REGRESSAO.R <==
################################################################################
# NOME: TABELAS DE REGRESSAO
# AUTOR: FILLIPE
# DESCRICAO: CRIACAO DAS TABELAS DE REGRESSAO CONSOLIDADAS.
###############################################################################
library(kableExtra)
library(tidyverse)
#################### GRUPO DE CONTROLE 1 #######################################

Controle1_2010= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle1_2010.RDs')
Controle1_2011= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle1_2011.RDs')
Controle1_2012= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle1_2012.RDs')
Controle1_2013= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle1_2013.RDs')

Controle1_2010_EF= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle1_2010_EF.RDs')
Controle1_2011_EF= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle1_2011_EF.RDs')
Controle1_2012_EF= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle1_2012_EF.RDs')
Controle1_2013_EF= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle1_2013_EF.RDs')

fator_elasticidade = readRDS('E:/salmin_reg2/R Sal Min/TABELAS - DESCRITIVAS/FATOR_ELASTICIDADE.RDs')

elasticidade = c(Controle1_2010$coefficients["trat_p"]*2.91, 
                 Controle1_2011$coefficients["trat_p"]*2.34, 
                 Controle1_2012$coefficients["trat_p"]*2.93,
                 Controle1_2013$coefficients["trat_p"]*2.32,
                 Controle1_2010_EF$coefficients["trat_p"]*2.91, 
                 Controle1_2011_EF$coefficients["trat_p"]*2.34, 
                 Controle1_2012_EF$coefficients["trat_p"]*2.93,
                 Controle1_2013_EF$coefficients["trat_p"]*2.32 )

elasticidade = unname(elasticidade)
elasticidade = round(elasticidade,3)
elasticidade = as.character(elasticidade)

extra_rows = tribble(
  ~term, ~a, ~b, ~c, ~d, ~e, ~f, ~g, ~h,
  'Efeitos Fixos de Firma', 'Não', 'Não', 'Não', 'Não', 'Sim', 'Sim', 'Sim', 'Sim',
  'Controle de Tamanho de Firma', 'Sim', 'Sim', 'Sim', 'Sim', 'Não', 'Não', 'Não', 'Não',
  'elasticidade', elasticidade[1], elasticidade[2], elasticidade[3], elasticidade[4], elasticidade[5], elasticidade[6], elasticidade[7], elasticidade[8]
)

A=modelsummary(models = list(
  '(2010)' = Controle1_2010, 
  '(2011)' = Controle1_2011, 
  '(2012)' = Controle1_2012, 
  '(2013)' = Controle1_2013, 
  '(2010)' = Controle1_2010_EF, 
  '(2011)' = Controle1_2011_EF, 
  '(2012)' = Controle1_2012_EF, 
  '(2013)' = Controle1_2013_EF
), 
gof_omit = c('^AIC|^BIC|^Log|^FE|^Std.|Pseudo|R2 Within'),
stars = c('*' = 0.1, '**' = 0.05, '***' = 0.01),
coef_omit = '^totemp1|^educa|^tam|^fx|^part|^industria|^genero|^servicos|^construcao|^comercio|^periodo|grupo_trat',
title = 'Resultados de Regressão',
escape = FALSE, 
add_rows = extra_rows,  
output = 'E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/CONTROLE1.docx'
)

#################### GRUPO DE CONTROLE 2 #######################################
rm(list=ls())
gc()
Controle2_2010= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle2_2010.RDs')
Controle2_2011= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle2_2011.RDs')
Controle2_2012= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle2_2012.RDs')
Controle2_2013= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle2_2013.RDs')

Controle2_2010_EF= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle2_2010_EF.RDs')
Controle2_2011_EF= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle2_2011_EF.RDs')
Controle2_2012_EF= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle2_2012_EF.RDs')
Controle2_2013_EF= readRDS('E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle2_2013_EF.RDs')

fator_elasticidade = readRDS('E:/salmin_reg2/R Sal Min/TABELAS - DESCRITIVAS/FATOR_ELASTICIDADE.RDs')

elasticidade = c(Controle2_2010$coefficients["trat_p"]*2.91, 
                 Controle2_2011$coefficients["trat_p"]*2.34, 
                 Controle2_2012$coefficients["trat_p"]*2.93,
                 Controle2_2013$coefficients["trat_p"]*2.32,
                 Controle2_2010_EF$coefficients["trat_p"]*2.91, 
                 Controle2_2011_EF$coefficients["trat_p"]*2.34, 
                 Controle2_2012_EF$coefficients["trat_p"]*2.93,
                 Controle2_2013_EF$coefficients["trat_p"]*2.32 )

elasticidade = unname(elasticidade)
elasticidade = round(elasticidade,3)
elasticidade = as.character(elasticidade)

extra_rows = tribble(
  ~term, ~a, ~b, ~c, ~d, ~e, ~f, ~g, ~h,
  'Efeitos Fixos de Firma', 'Não', 'Não', 'Não', 'Não', 'Sim', 'Sim', 'Sim', 'Sim',
  'Controle de Tamanho de Firma', 'Sim', 'Sim', 'Sim', 'Sim', 'Não', 'Não', 'Não', 'Não',
  'elasticidade', elasticidade[1], elasticidade[2], elasticidade[3], elasticidade[4], elasticidade[5], elasticidade[6], elasticidade[7], elasticidade[8]
)


A=modelsummary(models = list(
  '(2010)' = Controle2_2010, 
  '(2011)' = Controle2_2011, 
  '(2012)' = Controle2_2012, 
  '(2013)' = Controle2_2013, 
  '(2010)' = Controle2_2010_EF, 
  '(2011)' = Controle2_2011_EF, 
  '(2012)' = Controle2_2012_EF, 
  '(2013)' = Controle2_2013_EF
), 
gof_omit = c('^AIC|^BIC|^Log|^FE|^Std.|Pseudo|R2 Within'),
stars = c('*' = 0.1, '**' = 0.05, '***' = 0.01),
coef_omit = '^totemp1|^educa|^tam|^fx|^part|^industria|^genero|^servicos|^construcao|^comercio|^periodo|grupo_trat',
title = 'Resultados de Regressão',
escape = FALSE, 
add_rows = extra_rows,  
output = 'E:/salmin_reg2/R Sal Min/REGRESSÕES/GERAL/Controle2.docx'
)
